Move plotting from match-filter to plotting
